requirements.txt
bcrypt==3.1.1
cffi==1.9.1
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector-python-rf==2.2.1
pycparser==2.17
PyJWT==1.4.2
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.11
